Reorganized tests to run in a set order
