+++ BrickOrdering/src/test/resources/reply1.json
@@ -1,1 +1,4 @@
-3
+{
+    "orderReference": 2,
+    "numberOfBricks": 26
+}
