+++ BrickOrdering/src/test/resources/reply3.json
@@ -1,14 +1,1 @@
-[
-    {
-        "orderReference": 1,
-        "numberOfBricks": 50
-    },
-    {
-        "orderReference": 2,
-        "numberOfBricks": 26
-    },
-    {
-        "orderReference": 3,
-        "numberOfBricks": 33
-    }
-]
+4
