+++ BrickOrdering/src/test/resources/request4.json
@@ -1,0 +1,1 @@
+506
